requirements.txt
attrs==25.1.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.1
darkdetect==0.8.0
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
lxml==5.3.0
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
PyQt6==6.8.0
PyQt6-Qt6==6.8.1
PyQt6_sip==13.9.1
PySocks==1.7.1
python-docx==1.1.2
python-dotenv==1.0.1
requests==2.32.3
selenium==4.28.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
